Customer data
There are two files that are read in as input. These are items.text which include the customer grocery items and their associated ID's. The other is orders.text that include the customer ID, date, item number and associated quantity, and finally the last line contains the payment method line. This information is parsed and sorted in groceries.cpp.. A final list of orders is printed to the console when the program is run.